+++ Release Notes.md
@@ -10,4 +10,5 @@ This version is the first public beta release of TagKit.
 * `Slugifiable` is a protocol that describes slugifiable types and adds additional functionality to any types that implement it.
 * `SlugConfiguration` can be used to customized the slugified result.
 * `Taggable` is a protocol that describes taggable types and adds additional functionality to any types that implement it and collections that contain them.
+* `TagList` renders a collection of tags using a custom view builder.
 * `TagTextField` automatically slugifies any text that is entered into it.